app: extract sct,compactblock into chain crate
