+++ app/src/action_handler/actions/proposal/submit.rs
@@ -7,7 +7,7 @@ use decaf377::Fq;
 use once_cell::sync::Lazy;
 use rand_chacha::{rand_core::SeedableRng, ChaCha20Rng};
 
-use penumbra_chain::component::StateReadExt as _;
+use penumbra_chain::component::{StateReadExt as _, StateWriteExt as _};
 use penumbra_crypto::{
     keys::{FullViewingKey, NullifierKey},
     rdsa::{VerificationKey, VerificationKeyBytes},
@@ -20,9 +20,7 @@ use penumbra_transaction::{action::ProposalSubmit, Transaction};
 use penumbra_transaction::{AuthorizationData, WitnessData};
 
 use crate::action_handler::ActionHandler;
-use crate::compactblock::view::{StateReadExt as _, StateWriteExt as _};
 use crate::governance::{StateReadExt as _, StateWriteExt as _};
-use crate::sct::view::StateReadExt as _;
 use crate::shielded_pool::SupplyWrite;
 
 // IMPORTANT: these length limits are enforced by consensus! Changing them will change which
