+++ app/src/lib.rs
@@ -10,12 +10,10 @@ pub use mock_client::MockClient;
 pub use temp_storage_ext::TempStorageExt;
 
 pub mod app;
-pub mod compactblock;
 pub mod dao;
 pub mod dex;
 pub mod governance;
 pub mod ibc;
-pub mod sct;
 pub mod shielded_pool;
 pub mod stake;
 pub mod stubdex;
