--- app/src/sct/mod.rs
@@ -1,6 +1,0 @@
-// Like the DAO, this isn't a "component" because it doesn't do any execution,
-// but we want to create a collection of state that can be accessed from upstream
-// components and is cleanly separated
-
-pub mod state_key;
-pub mod view;
