+++ app/src/shielded_pool/note_manager.rs
@@ -1,9 +1,8 @@
 use super::{event, state_key, SupplyWrite};
-use crate::compactblock::view::{StateReadExt as _, StateWriteExt as _};
-use crate::sct::view::{StateReadExt as _, StateWriteExt as _};
 use anyhow::Result;
 use async_trait::async_trait;
 use penumbra_chain::{
+    component::{StateReadExt as _, StateWriteExt as _},
     sync::{StatePayload, StatePayloadDebugKind},
     NoteSource, SpendInfo,
 };
