+++ chain/src/component/view.rs
@@ -2,15 +2,16 @@ use std::str::FromStr;
 
 use anyhow::{anyhow, Context, Result};
 use async_trait::async_trait;
-#[cfg(feature = "penumbra-storage")]
 use penumbra_proto::{StateReadProto, StateWriteProto};
-#[cfg(feature = "penumbra-storage")]
 use penumbra_storage::{StateRead, StateWrite};
+use penumbra_tct as tct;
 use tendermint::Time;
 
 use crate::{
     params::{ChainParameters, FmdParameters},
-    state_key, Epoch,
+    state_key,
+    sync::CompactBlock,
+    Epoch,
 };
 
 /// This trait provides read access to common parts of the Penumbra
@@ -153,6 +154,53 @@ pub trait StateReadExt: StateRead {
             .await?
             .unwrap_or_default())
     }
+
+    // formerly compact block methods
+
+    async fn compact_block(&self, height: u64) -> Result<Option<CompactBlock>> {
+        self.get(&state_key::compact_block(height)).await
+    }
+
+    fn stub_compact_block(&self) -> CompactBlock {
+        self.object_get(state_key::stub_compact_block())
+            .unwrap_or_default()
+    }
+
+    // formerly sct methods
+
+    async fn stub_state_commitment_tree(&self) -> tct::Tree {
+        match self
+            .nonconsensus_get_raw(state_key::stub_state_commitment_tree().as_bytes())
+            .await
+            .unwrap()
+        {
+            Some(bytes) => bincode::deserialize(&bytes).unwrap(),
+            None => tct::Tree::new(),
+        }
+    }
+
+    async fn anchor_by_height(&self, height: u64) -> Result<Option<tct::Root>> {
+        self.get(&state_key::anchor_by_height(height)).await
+    }
+
+    async fn check_claimed_anchor(&self, anchor: tct::Root) -> Result<()> {
+        if anchor.is_empty() {
+            return Ok(());
+        }
+
+        if let Some(anchor_height) = self
+            .get_proto::<u64>(&state_key::anchor_lookup(anchor))
+            .await?
+        {
+            tracing::debug!(?anchor, ?anchor_height, "anchor is valid");
+            Ok(())
+        } else {
+            Err(anyhow!(
+                "provided anchor {} is not a valid SCT root",
+                anchor
+            ))
+        }
+    }
 }
 
 impl<T: StateRead + ?Sized> StateReadExt for T {}
@@ -216,6 +264,73 @@ pub trait StateWriteExt: StateWrite {
     // Signals that the epoch should end this block.
     fn signal_end_epoch(&mut self) {
         self.object_put(state_key::end_epoch_early(), true)
+    }
+
+    // formerly compact block methods
+
+    fn stub_put_compact_block(&mut self, compact_block: CompactBlock) {
+        self.object_put(state_key::stub_compact_block(), compact_block);
+    }
+
+    fn set_compact_block(&mut self, compact_block: CompactBlock) {
+        let height = compact_block.height;
+        self.put(state_key::compact_block(height), compact_block);
+    }
+
+    async fn height(&self) -> u64 {
+        self.get_block_height()
+            .await
+            .expect("block height must be set")
+    }
+
+    fn stub_put_state_commitment_tree(&mut self, tree: &tct::Tree) {
+        let bytes = bincode::serialize(&tree).unwrap();
+        self.nonconsensus_put_raw(
+            state_key::stub_state_commitment_tree().as_bytes().to_vec(),
+            bytes,
+        );
+    }
+
+    // formerly sct methods
+
+    fn set_sct_anchor(&mut self, height: u64, sct_anchor: tct::Root) {
+        tracing::debug!(?height, ?sct_anchor, "writing anchor");
+
+        self.put(state_key::anchor_by_height(height), sct_anchor);
+        self.put_proto(state_key::anchor_lookup(sct_anchor), height);
+    }
+
+    fn set_sct_block_anchor(&mut self, height: u64, sct_block_anchor: tct::builder::block::Root) {
+        tracing::debug!(?height, ?sct_block_anchor, "writing block anchor");
+
+        self.put(state_key::block_anchor_by_height(height), sct_block_anchor);
+        self.put_proto(state_key::block_anchor_lookup(sct_block_anchor), height);
+    }
+
+    fn set_sct_epoch_anchor(&mut self, index: u64, sct_block_anchor: tct::builder::epoch::Root) {
+        tracing::debug!(?index, ?sct_block_anchor, "writing epoch anchor");
+
+        self.put(state_key::epoch_anchor_by_index(index), sct_block_anchor);
+        self.put_proto(state_key::epoch_anchor_lookup(sct_block_anchor), index);
+    }
+
+    async fn write_sct(
+        &mut self,
+        height: u64,
+        sct: tct::Tree,
+        block_root: tct::builder::block::Root,
+        epoch_root: Option<tct::builder::epoch::Root>,
+    ) {
+        self.set_sct_anchor(height, sct.root());
+
+        self.set_sct_block_anchor(height, block_root);
+
+        if let Some(epoch_root) = epoch_root {
+            let index = self.epoch().await.expect("epoch must be set").index;
+            self.set_sct_epoch_anchor(index, epoch_root);
+        }
+
+        self.stub_put_state_commitment_tree(&sct);
     }
 }
