+++ view/src/worker.rs
@@ -349,7 +349,7 @@ async fn sct_divergence_check(
     height: u64,
     actual_root: penumbra_tct::Root,
 ) -> anyhow::Result<()> {
-    use penumbra_app::sct::state_key as sct_state_key;
+    use penumbra_chain::state_key as sct_state_key;
 
     let value = client
         .key_value(penumbra_proto::client::v1alpha1::KeyValueRequest {
